refactor: ErrorResponse 구현
