+++ src/main/java/kitchenpos/exception/ControllerAdvice.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,15 +9,15 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class ControllerAdvice {
 
     @ExceptionHandler(KitchenException.class)
-    public ResponseEntity<String> handleKitchenException(KitchenException e) {
+    public ResponseEntity<ErrorResponse> handleKitchenException(KitchenException e) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-            .body(e.getMessage());
+            .body(new ErrorResponse(e, e.getMessage()));
     }
 
     @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleRunTimeException(RuntimeException e) {
+    public ResponseEntity<ErrorResponse> handleRunTimeException(RuntimeException e) {
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-            .body(e.getMessage());
+            .body(new ErrorResponse(e, e.getMessage()));
     }
 
 }
